test(storage): move multi-check list tests into session_test
